refactor(llm): split providers module into per-class files
The single providers.py held the LLMProvider base, the LLMResponse
dataclass, and three concrete providers. The TS rule "default export
for class files (one class per file)" applies in spirit to Python
adapter packages too — each concrete provider has a different auth
shape, base URL, and capability flag. Splitting:

- providers/base.py — LLMProvider + LLMResponse
- providers/github.py — GitHubProvider
- providers/mistral.py — MistralProvider
- providers/groq.py — GroqProvider

Namespace package (PEP 420) — no __init__.py per the project rule.
provider_chain and the test file import each class from its module
directly.
